redo mtu and ntp tests correctly. Draft version
